refactor: created util folder to store current util.py as __init__.py, and moved dataset2pmtiles under util package
